# docs/swagger.yaml
definitions:
  model.BuyProductBody:
    properties:
      product_id:
        type: integer
      user_id:
        type: integer
    type: object
  model.LoginUserRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  model.UserBody:
    properties:
      balance:
        type: number
      email:
        type: string
      name:
        type: string
      password:
        type: string
      surname:
        type: string
      username:
        type: string
    required:
    - email
    - name
    - surname
    type: object
info:
  contact: {}
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Login Users
      parameters:
      - description: Request of Creating User Object
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Login User
      tags:
      - Auth
  /products/all:
    get:
      consumes:
      - application/json
      description: Getting products with given request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Getting products
      tags:
      - Products
  /products/buy:
    post:
      consumes:
      - application/json
      description: Buying product with given request
      parameters:
      - description: Request of Buying Product Object
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.BuyProductBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Buying product
      tags:
      - Products
  /users:
    post:
      consumes:
      - application/json
      description: Creating User with given request
      parameters:
      - description: Request of Creating User Object
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UserBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Creating User
      tags:
      - Users
  /users/all:
    get:
      consumes:
      - application/json
      description: Getting Users with given request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Getting Users
      tags:
      - Users
swagger: "2.0"